Prevent the OSStatus -1712 timeout errors
I can't find details on this, but the default AppleEvent timeout value
was never a problem until now. It's either because the plugin used to
work better and faster for users or there has been a change in macOS.
Either way we remove the timeout for AppleEvents here, so long running
operations should no longer trigger a timeout error, including
displaying an alert for an user who is away from their computer.