Update the error when document is not found to disable Rosetta
We used to tell people to enable Rosetta with ARM Macs and macOS 11.3 <
due to an AppleScript messaging bug between Zotero in Rosetta and
Word running in ARM, so some of these users may still be running
it that way because of us.

Also remove an old irrelevant error message about Apple migration
manager breaking Word migration in 2013.